SAM-66238--FE--APP--AI-SCRIBE - extract to components
