+++ vim--ai-scribe-app/src/components/organisms/ai-scribe-demo/Note.interface.ts
@@ -1,0 +1,12 @@
+
+export interface Note {
+  id: string;
+  patientName: string;
+  timestamp: string;
+  soap: {
+    subjective: string;
+    objective: string;
+    assessment: string;
+    plan: string;
+  };
+}
