+++ vim--ai-scribe-app/src/components/organisms/ai-scribe-demo/formatTime.util.ts
@@ -1,0 +1,7 @@
+export const formatTime = (seconds: number) => {
+  const mins = Math.floor(seconds / 60);
+  const secs = seconds % 60;
+  return `${mins.toString().padStart(2, "0")}:${secs
+    .toString()
+    .padStart(2, "0")}`;
+};
