+++ vim--ai-scribe-app/src/components/organisms/ai-scribe-demo/keywords.mock.ts
