+++ vim--ai-scribe-app/src/components/organisms/ai-scribe-demo/transcription.mock.ts
